pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-apps"
version = "0.1.0"
description = "Runbook utilities for the cloud-apps stack"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ansible==13.4.0",
    "from-root==1.3.0",
    "pydantic==2.10.6",
    "python-dotenv==1.2.1",
    "PyYAML==6.0.3",
    "requests==2.32.5",
    "ruff==0.15.5",
    "types-requests==2.32.4.20260107",
]

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["runbook*", "src*"]
